Change size_t function args/returns to int in Base.h
AllocZero/AllocArray/AllocStruct, memdup, memeq, MurmurHash2 and
RoundToPowerOf2 now use int for sizes/lengths (matching the int-length
convention elsewhere), and the redundant size_t RoundUp overload is removed
in favor of the int one. Callers that hold size_t add an (int) cast at the
boundary (Archive, EngineMupdf, MobiDoc, ChmFile, Dict, GuessFileType);
RoundToPowerOf2 returns -1 on overflow. Update Base_ut/SumatraUnitTests.

prompt: change uses of size_t as function argument or return value to int in Base.h, update/simplify callers, and remove redundant casts in the modified code

+++ src/ChmFile.cpp
@@ -65,7 +65,7 @@ Str ChmFile::GetData(Str fileName) const {
     }
 
     // +1 for 0 terminator for C string compatibility
-    u8* d = AllocArray<u8>(len + 1);
+    u8* d = AllocArray<u8>((int)(len + 1));
     if (!d) {
         return {};
     }
